Real-time update on index meta data
