+++ log/poll.log
@@ -1,26 +1,13 @@
 [INFO] [20/6/2015] GET /pf35v/results
 [INFO] [20/6/2015] GET /pf35v/json-results
+[INFO] [20/6/2015] GET /pf35v/img/favicon.ico
+[INFO] [20/6/2015] GET /img/favicon.ico
 [INFO] [20/6/2015] PUT /pf35v
-[INFO] [20/6/2015] GET /pf35v/results
-[INFO] [20/6/2015] GET /pf35v/json-results
-[INFO] [20/6/2015] PUT /pf35v
-[WARNING] [20/6/2015] Attempted second vote on poll: pf35v
 [INFO] [20/6/2015] GET /pf35v/results
 [INFO] [20/6/2015] GET /pf35v/json-results
 [INFO] [20/6/2015] GET /
-[INFO] [20/6/2015] GET /create
-[INFO] [20/6/2015] From the polls/create route
-[INFO] [20/6/2015] GET /pf35v
-[INFO] [20/6/2015] PUT /pf35v
-[WARNING] [20/6/2015] Attempted second vote on poll: pf35v
-[INFO] [20/6/2015] GET /pf35v/results
-[INFO] [20/6/2015] GET /pf35v/json-results
-[INFO] [20/6/2015] GET /create
-[INFO] [20/6/2015] From the polls/create route
-[INFO] [20/6/2015] GET /create
-[INFO] [20/6/2015] From the polls/create route
-[INFO] [20/6/2015] GET /img/favicon.ico
-[INFO] [20/6/2015] GET /create
-[INFO] [20/6/2015] From the polls/create route
-[INFO] [20/6/2015] GET /img/favicon.ico
-[INFO] [20/6/2015] GET /
+[INFO] [20/6/2015] GET /8354y
+[INFO] [20/6/2015] PUT /8354y
+[INFO] [20/6/2015] GET /8354y/results
+[INFO] [20/6/2015] GET /8354y/json-results
+[INFO] [20/6/2015] GET /8354y/img/favicon.ico
